refactor: move renderer-dependent utils into workbench scope
- extract renderer-only isPrimitiveNode into src/renderer/utils/nodeTypeGuards.ts
- relocate nodeHelpUtil, modelMetadataUtil, nodeDefOrderingUtil to src/workbench/utils/
- update services, stores, scripts, and tests to use the new paths
- rerun layer analyzer to confirm fewer base→renderer/workbench edges
