+++ src/services/nodeHelpService.ts
@@ -1,7 +1,7 @@
 import { api } from '@/scripts/api'
 import type { ComfyNodeDefImpl } from '@/stores/nodeDefStore'
 import { NodeSourceType, getNodeSource } from '@/types/nodeSource'
-import { extractCustomNodeName } from '@/utils/nodeHelpUtil'
+import { extractCustomNodeName } from '@/workbench/utils/nodeHelpUtil'
 
 class NodeHelpService {
   async fetchNodeHelp(node: ComfyNodeDefImpl, locale: string): Promise<string> {
