+++ src/utils/typeGuardUtil.ts
@@ -1,16 +1,9 @@
-import type { PrimitiveNode } from '@/extensions/core/widgetInputs'
 import type {
   INodeSlot,
   LGraph,
   LGraphNode,
   Subgraph
 } from '@/lib/litegraph/src/litegraph'
-
-export function isPrimitiveNode(
-  node: LGraphNode
-): node is PrimitiveNode & LGraphNode {
-  return node.type === 'PrimitiveNode'
-}
 
 /**
  * Check if an error is an AbortError triggered by `AbortController#abort`
