+++ tests-ui/tests/litegraph/core/LGraphNode.widgetOrder.test.ts
@@ -2,7 +2,7 @@ import { beforeEach, describe, expect, it } from 'vitest'
 
 import { LGraphNode } from '@/lib/litegraph/src/litegraph'
 import type { ISerialisedNode } from '@/lib/litegraph/src/types/serialisation'
-import { sortWidgetValuesByInputOrder } from '@/utils/nodeDefOrderingUtil'
+import { sortWidgetValuesByInputOrder } from '@/workbench/utils/nodeDefOrderingUtil'
 
 describe('LGraphNode widget ordering', () => {
   let node: LGraphNode
